Face_Recognition/requirements.txt
opencv-python
facenet-pytorch
torch
torchaudio
torchvision
boto3
pillow


